requirements.txt
-q google-generativeai
flask
pypdf2
python-docx
exceptions
re
ast
pandas
pymongo